Added comments
-Changed some method parameters
-deleted some compilation warnings

+++ brickBreak/BallBrick.cpp
@@ -10,7 +10,7 @@ bool BallBrick::isDestroyed(std::vector<Ball>& ref, std::vector<Particle>& Parti
 	Brick::isDestroyed(ref, ParticleVector);
 	if (hits==0) {
 		/*Send the ball created in a random direction*/
-		float angle=(rand()%360)*3.14/360.0f;
+		float angle= (float)((rand()%360)*3.14/360.0f);
 		sf::Vector2f speedDir(15 * cos(angle), 15 * sin(angle));
 		ref.push_back(Ball(sf::Vector2f((pos.x + dim.x / 2) - 10, (pos.y + dim.y / 2) - 10), 10, speedDir, sf::Color::Red));
 		return true;
